fix: Correct copyright headers of files forked from TVL
This fixes the copyright headers of files that were forked from the
TVL depot and not modified substantially enough to warrant
relicensing.

Note that there's also some code under //third_party which is from TVL
and doesn't include the copyright notes required by TVL, but I didn't
touch those in this commit.

This commit is exempt from CLA as it only modifies files to which we
hold the copyright.